+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master
 
+* [bugfix] Use childprocess to start ssh related commands. Fixes tty issue
+  when using `shelly console` on JRuby
 * [bugfix] Capture password without echo only if terminal is present
 
 ## 0.4.28 / 2014-03-25